+++ TDEngine2/include/ecs/ICameraSystem.h
@@ -20,6 +20,7 @@ namespace TDEngine2
 	class IPerspectiveCamera;
 	class IOrthoCamera;
 	class IRenderer;
+	class IDebugUtility;
 
 
 	/*!
@@ -75,6 +76,10 @@ namespace TDEngine2
 			*/
 
 			TDE2_API virtual E_RESULT_CODE SetMainCamera(const ICamera* pCamera) = 0;
+
+#if TDE2_EDITORS_ENABLED
+			TDE2_API virtual E_RESULT_CODE SetDebugUtility(IDebugUtility* pDebugUtility) = 0;
+#endif
 
 			/*!
 				\brief The method returns a pointer to a main camera from which the scene is rendered